requirements.txt
flask==1.0.2
flask-cors==3.0.7
flask-sqlalchemy==2.1
google-auth==1.5.1
requests==2.21.0
